+++ server/src/test/documentFormating.test.ts
@@ -55,7 +55,7 @@ const getEdits = async (document: TextDocument) => {
     {
       textDocument,
       options: {
-        tabSize: 2,
+        tabSize: 4,
         insertSpaces: false,
         trimTrailingWhitespace: true,
       },
@@ -87,6 +87,13 @@ describe("Document formating", () => {
       const newText = await getNewText(documentText);
       expect(newText).toEqual("/ { };");
     });
+
+    test("new line and tab state of doc", async () => {
+      const documentText = `\n\tn1: node {\n\t\tprop1;\n\t\tprop2;\n\t};`;
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual(`n1: node {\n\tprop1;\n\tprop2;\n};`);
+    });
+
     test("Node extra new line from top", async () => {
       const documentText = "\n/ { };";
       const newText = await getNewText(documentText);
@@ -168,12 +175,12 @@ describe("Document formating", () => {
     test("Comment before ;", async () => {
       const documentText = "/ {\n} /* abc */  ;";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ { }; /* abc */");
+      expect(newText).toEqual("/ { };\t/* abc */");
     });
     test("Comments before ;", async () => {
       const documentText = "/ {\n} /* abc1 */     /* abc2 */   ;";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ { }; /* abc1 */ /* abc2 */");
+      expect(newText).toEqual("/ { };\t/* abc1 */\t/* abc2 */");
     });
   });
 
@@ -204,6 +211,13 @@ describe("Document formating", () => {
       const newText = await getNewText(documentText);
       expect(newText).toEqual("/ {\n\tnode@20 { };\n};");
     });
+
+    test("Node address with unknown value", async () => {
+      const documentText = "/{node1@FOO_BAR{};};";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual("/ {\n\tnode1@FOO_BAR { };\n};");
+    });
+
     test("Multiple space between name and { no address", async () => {
       const documentText = "/ {\n\tnode  { };\n};";
       const newText = await getNewText(documentText);
@@ -284,15 +298,15 @@ describe("Document formating", () => {
       const newText = await getNewText(documentText);
       expect(newText).toEqual("/ {\n\tnode { };\n};");
     });
-    test("Comment before ;", async () => {
+    test("Comment before ; - case 1", async () => {
       const documentText = "/ {\n\tnode { } /* abc */   ;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tnode { }; /* abc */\n};");
+      expect(newText).toEqual("/ {\n\tnode { };\t/* abc */\n};");
     });
-    test("Comments before ;", async () => {
-      const documentText = "/ {\n\tnode { } /* abc1 */ /* abc2 */   ;\n};";
+    test("Comments before ; - case 2", async () => {
+      const documentText = "/ {\n\tnode { }/* abc1 */ /* abc2 */   ;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tnode { }; /* abc1 */ /* abc2 */\n};");
+      expect(newText).toEqual("/ {\n\tnode { };\t/* abc1 */\t/* abc2 */\n};");
     });
   });
 
@@ -381,12 +395,12 @@ describe("Document formating", () => {
     test("Comment before ;", async () => {
       const documentText = "&n1 {\n} /* abc */  ;";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("&n1 { }; /* abc */");
+      expect(newText).toEqual("&n1 { };\t/* abc */");
     });
     test("Comments before ;", async () => {
       const documentText = "&n1 {\n} /* abc1 */     /* abc2 */   ;";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("&n1 { }; /* abc1 */ /* abc2 */");
+      expect(newText).toEqual("&n1 { };\t/* abc1 */\t/* abc2 */");
     });
   });
 
@@ -470,13 +484,13 @@ describe("Document formating", () => {
     test("Single comment between path and ;", async () => {
       const documentText = "/delete-node/ &n1 /* abc */  ;";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/delete-node/ &n1; /* abc */");
+      expect(newText).toEqual("/delete-node/ &n1;\t/* abc */");
     });
 
     test("Multiple comments between path and ;", async () => {
       const documentText = "/delete-node/ &n1 /* abc1 */   /* abc2 */  ;";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/delete-node/ &n1; /* abc1 */ /* abc2 */");
+      expect(newText).toEqual("/delete-node/ &n1;\t/* abc1 */\t/* abc2 */");
     });
 
     test("One comment after ; and multiple comments between path and ;", async () => {
@@ -484,12 +498,24 @@ describe("Document formating", () => {
         "/delete-node/ &n1 /* abc1 */   /* abc2 */  ; /* abc3 */";
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
-        "/delete-node/ &n1; /* abc1 */ /* abc2 */ /* abc3 */"
+        "/delete-node/ &n1;\t/* abc1 */\t/* abc2 */\t/* abc3 */"
       );
     });
   });
 
   describe("Line Comment", () => {
+    test("new line on top of document", async () => {
+      const documentText = "\n// foo";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual("// foo");
+    });
+
+    test("multiple new lines on top of document", async () => {
+      const documentText = "\n\n// foo";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual("// foo");
+    });
+
     test("Correct indentation in level 1", async () => {
       const documentText = "/ {\n// foo\n\tnode { };\n};";
       const newText = await getNewText(documentText);
@@ -516,6 +542,17 @@ describe("Document formating", () => {
   });
 
   describe("Block Comment", () => {
+    test("new line on top of document", async () => {
+      const documentText = "\n/* foo */";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual("/* foo */");
+    });
+
+    test("multiple new lines on top of document", async () => {
+      const documentText = "\n\n/* foo */";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual("/* foo */");
+    });
     test("Correct indentation in level 1 single line", async () => {
       const documentText = "/ {\n/* foo */\n\tnode { };\n};";
       const newText = await getNewText(documentText);
@@ -541,6 +578,21 @@ describe("Document formating", () => {
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
         "/ {\n\tnode {\n\t\t/* foo\n\t\t * bar\n\t\t */\n\t};\n};"
+      );
+    });
+
+    test("in ref node", async () => {
+      const documentText = "&n1 {\n\tprop1;\n\t/* foo */\n\tnode { };\n};";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual("&n1 {\n\tprop1;\n\t/* foo */\n\tnode { };\n};");
+    });
+
+    test("in node", async () => {
+      const documentText =
+        "/ {\n\t\t/* foo */\nnode {\n\tprop1;\n\t/* foo */\n\tnode { };\n};\n}";
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual(
+        "/ {\n\t/* foo */\n\tnode {\n\t\tprop1;\n\t\t/* foo */\n\t\tnode { };\n\t};\n}"
       );
     });
   });
@@ -593,7 +645,7 @@ describe("Document formating", () => {
     test("Single comment between path and ;", async () => {
       const documentText = "/ {\n\t/delete-property/ n1 /* abc */;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\t/delete-property/ n1; /* abc */\n};");
+      expect(newText).toEqual("/ {\n\t/delete-property/ n1;\t/* abc */\n};");
     });
 
     test("Multiple comments between path and ;", async () => {
@@ -601,7 +653,7 @@ describe("Document formating", () => {
         "/ {\n\t/delete-property/ n1 /* abc1 */  /* abc2 */;\n};";
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
-        "/ {\n\t/delete-property/ n1; /* abc1 */ /* abc2 */\n};"
+        "/ {\n\t/delete-property/ n1;\t/* abc1 */\t/* abc2 */\n};"
       );
     });
 
@@ -610,7 +662,7 @@ describe("Document formating", () => {
         "/ {\n\t/delete-property/ n1 /* abc1 */  /* abc2 */;   /* abc3 */\n};";
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
-        "/ {\n\t/delete-property/ n1; /* abc1 */ /* abc2 */ /* abc3 */\n};"
+        "/ {\n\t/delete-property/ n1;\t/* abc1 */\t/* abc2 */\t/* abc3 */\n};"
       );
     });
   });
@@ -702,13 +754,29 @@ describe("Document formating", () => {
     test("comment between array value", async () => {
       const documentText = "/ {\n\tprop1 =   <10  /* foo */  20>;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = <10 /* foo */ 20>;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <10\t/* foo */ 20>;\n};");
     });
     test("single new line between array value", async () => {
       const documentText = "/ {\n\tprop1 =   <10\n20>;\n};";
       const newText = await getNewText(documentText);
       expect(newText).toEqual("/ {\n\tprop1 = <10\n\t\t\t20>;\n};");
     });
+
+    test("Multi line string", async () => {
+      const documentText = '/ {\n\tprop1 = "FOO\nBAR";\n};';
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual('/ {\n\tprop1 = "FOO\nBAR";\n};');
+    });
+
+    test("Escaped string", async () => {
+      const documentText =
+        '/ {\n\tval = "XA\nXPLUS\nXB", "XSTR1 \\" plus \\" XSTR2";\n\tprop = <10>;\n};';
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual(
+        '/ {\n\tval = "XA\nXPLUS\nXB", "XSTR1 \\" plus \\" XSTR2";\n\tprop = <10>;\n};'
+      );
+    });
+
     test("multiple new lines between array value", async () => {
       const documentText = "/ {\n\tprop1 =   <10\n\n20>;\n};";
       const newText = await getNewText(documentText);
@@ -717,22 +785,22 @@ describe("Document formating", () => {
     test("comment and single new line between array value", async () => {
       const documentText = "/ {\n\tprop1 = <10  /* foo */\n20>;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = <10 /* foo */\n\t\t\t20>;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <10\t/* foo */\n\t\t\t20>;\n};");
     });
     test("comment and multiple new lines between array value", async () => {
-      const documentText = "/ {\n\tprop1 = <10  /* foo */\n\n20>;\n};";
+      const documentText = "/ {\n\tprop1 = <10/* foo */\n\n20>;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = <10 /* foo */\n\t\t\t20>;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <10\t/* foo */\n\t\t\t20>;\n};");
     });
     test("comment after <", async () => {
       const documentText = "/ {\n\tprop1 = <    /* foo */    10>;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = < /* foo */ 10>;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <\t/* foo */ 10>;\n};");
     });
     test("comment before >", async () => {
       const documentText = "/ {\n\tprop1 = <10   /* foo */      >;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = <10 /* foo */ >;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <10\t/* foo */ >;\n};");
     });
     test("multiple space before >", async () => {
       const documentText = "/ {\n\tprop1 = <10   >;\n};";
@@ -767,12 +835,12 @@ describe("Document formating", () => {
     test("comment after comma on new line", async () => {
       const documentText = "/ {\n\tprop1 = <10>,\n/* foo */<20>;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = <10>, /* foo */ <20>;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <10>,\t/* foo */ <20>;\n};");
     });
     test("comment before comma", async () => {
       const documentText = "/ {\n\tprop1 = <10>   /* foo */, <20>;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = <10>, /* foo */ <20>;\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = <10>,\t/* foo */ <20>;\n};");
     });
     test("multiple before ;", async () => {
       const documentText = "/ {\n\tprop1 = <10>   ;\n};";
@@ -811,13 +879,13 @@ describe("Document formating", () => {
     test("Single comment between path and ;", async () => {
       const documentText = "/ {\n\tprop1 /* abc */;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1; /* abc */\n};");
+      expect(newText).toEqual("/ {\n\tprop1;\t/* abc */\n};");
     });
 
     test("Multiple comments between path and ;", async () => {
       const documentText = "/ {\n\tprop1 /* abc1 */  /* abc2 */;\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1; /* abc1 */ /* abc2 */\n};");
+      expect(newText).toEqual("/ {\n\tprop1;\t/* abc1 */\t/* abc2 */\n};");
     });
 
     test("One comment after ; and multiple comments between path and ;", async () => {
@@ -825,7 +893,7 @@ describe("Document formating", () => {
         "/ {\n\tprop1 /* abc1 */  /* abc2 */;   /* abc3 */\n};";
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
-        "/ {\n\tprop1; /* abc1 */ /* abc2 */ /* abc3 */\n};"
+        "/ {\n\tprop1;\t/* abc1 */\t/* abc2 */\t/* abc3 */\n};"
       );
     });
 
@@ -880,7 +948,7 @@ describe("Document formating", () => {
     test("CMacroCall assign param macro before ,", async () => {
       const documentText = "/ {\n\tprop1 = ADD (10     /* foo */    , 20);\n};";
       const newText = await getNewText(documentText);
-      expect(newText).toEqual("/ {\n\tprop1 = ADD(10, /* foo */ 20);\n};");
+      expect(newText).toEqual("/ {\n\tprop1 = ADD(10,\t/* foo */ 20);\n};");
     });
 
     test("Complex Expression extra stapce after (", async () => {
@@ -915,10 +983,10 @@ describe("Document formating", () => {
       expect(newText).toEqual("/ {\n\tprop1 = [10 20 30 40];\n};");
     });
     test("byte string allow on new line", async () => {
-      const documentText = "/ {\n\tprop1 = [10\n20\n30\n40]   ;\n};";
+      const documentText = "/ {\n\tprop11 = [10\n20\n30\n40]   ;\n};";
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
-        "/ {\n\tprop1 = [10\n\t\t\t20\n\t\t\t30\n\t\t\t40];\n};"
+        "/ {\n\tprop11 = [10\n\t\t\t 20\n\t\t\t 30\n\t\t\t 40];\n};"
       );
     });
   });
@@ -930,10 +998,42 @@ describe("Document formating", () => {
       expect(newText).toEqual("{ };");
     });
     test("Clean when remove new lines", async () => {
-      const documentText = "{\n\tprop1 = <10>, \n<20>,            \n<30>;\n};";
+      const documentText = "{\n\tprop11 = <10>, \n<20>,            \n<30>;\n};";
       const newText = await getNewText(documentText);
       expect(newText).toEqual(
-        "{\n\tprop1 = <10>,\n\t\t\t<20>,\n\t\t\t<30>;\n};"
+        "{\n\tprop11 = <10>,\n\t\t\t <20>,\n\t\t\t <30>;\n};"
+      );
+    });
+
+    test("move to new lines", async () => {
+      const documentText = '&spi1_nss_pa4 { slew-rate = "very-high-speed"; };';
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual(
+        '&spi1_nss_pa4 {\n\tslew-rate = "very-high-speed";\n};'
+      );
+    });
+
+    test("empty node", async () => {
+      const documentText = `arduino_i2c: &i2c1 {};\narduino_spi: &spi1 {};`;
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual(
+        `arduino_i2c: &i2c1 { };\narduino_spi: &spi1 { };`
+      );
+    });
+
+    test("move to new lines", async () => {
+      const documentText = `sysclk: system-clock {
+	compatible = "fixed-clock";
+	clock-frequency = <25000000>;
+	#clock-cells = <0>;
+};`;
+      const newText = await getNewText(documentText);
+      expect(newText).toEqual(
+        `sysclk: system-clock {
+	compatible = "fixed-clock";
+	clock-frequency = <25000000>;
+	#clock-cells = <0>;
+};`
       );
     });
   });