+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.4.5] - 2025-07-02
 
+### Added
+
+- Support for node to be refereance by node path e.g
+
+```devicetree
+&{/node1/node2@20/node3}{
+  prop1;
+}
+```
+
 ### Fixed
 
 - Fix issue where `/delete-node/` did not respect order when used inside child node
@@ -25,6 +35,38 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 n1& {
     /delete-node/ node2;
     node2 { };
+};
+```
+
+- Parser can handel node names starting with a number. A diagnostic error is show in this case.
+- Parser can handel node unit addresses starting with `0x`. A diagnostic warning is show in this case.
+- Parser can handel node unit addresses ending with `ULL`. A diagnostic warning is show in this case.
+- Parser can handel node unit addresses with `_` e.g `node@8_000_00`.
+- Semantic tokens for node paths e.g. `...&{/node1/node2@20/node3}`
+- Diagnostic error when node path referance have space between ampersand and open curly e.g. `&  {...}`
+- Comment formating inside a ref node e.g.
+
+```devicetree
+&n1 {
+  /* foo */
+  prop1;
+}
+```
+
+- Formating when properties are included e.g.
+
+```devicetree
+&qspi {
+	nrf70: nrf7002@1 {
+		compatible = "nordic,nrf7002-qspi";
+		status = "okay";
+		reg = <1>;
+		qspi-frequency = <24000000>;
+		qspi-quad-mode;
+
+		#include "nrf70_common.dtsi"
+		#include "nrf70_common_5g.dtsi"
+	};
 };
 ```